[site-isolation] Add touch event region to EventRegion https://bugs.webkit.org/show_bug.cgi?id=287916 rdar://145101984
Reviewed by Simon Fraser.

For the first part of rdar://133867737, add touch event regions to EventRegion and populate it based
on the RenderStyle value of eventListenerRegionTypes, as we do for wheel events on Mac.

* Source/WTF/wtf/PlatformEnable.h:
* Source/WTF/wtf/PlatformEnableCocoa.h:
* Source/WebCore/rendering/EventRegion.cpp:
(WebCore::EventRegion::EventRegion):
(WebCore::EventRegion::uniteEventListeners):
(WebCore::EventRegion::eventListenerRegionTypesForPoint const):
(WebCore::EventRegion::eventListenerRegionForType const):
* Source/WebCore/rendering/EventRegion.h:
* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::patternForEventListenerRegionType):
* Source/WebCore/rendering/style/RenderStyleConstants.cpp:
(WebCore::operator<<):
* Source/WebCore/rendering/style/RenderStyleConstants.h:
* Source/WebCore/style/StyleAdjuster.cpp:
(WebCore::Style::Adjuster::computeEventListenerRegionTypes):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:

Canonical link: https://commits.webkit.org/290717@main

+++ Source/WebCore/page/scrolling/ScrollingTree.h
@@ -58,7 +58,7 @@ class ScrollingTreeOverflowScrollProxyNode;
 class ScrollingTreePositionedNode;
 class ScrollingTreeScrollingNode;
 class ScrollingTreeFrameHostingNode;
-enum class EventListenerRegionType : uint8_t;
+enum class EventListenerRegionType : uint16_t;
 
 using FramesPerSecond = unsigned;
 using PlatformDisplayID = uint32_t;
